A new cluster-aware indexer for the MAP
Previously we were using a regular ArchivesSpace indexer to read
records from the ArchivesSpace backend and POST them to MAP's Solr
instance.  This won't work in production because we'll have N
ArchivesSpace instances and M MAP instances and we don't want every
ArchivesSpace instance to have to know about every MAP instance's
Solr.

This indexer is much like the ArchivesSpace periodic indexer (in that
it wakes up periodically, figures out what's new, pulls out the
records and maps them), but it writes to a table in the MAP database
instead of directly to Solr.  That way, MAP instances can consume a
stream of record updates from this new table to bring themselves
up-to-date.

Mapping records to Solr documents is done here, rather than on the MAP
side, because it seems likely that it'll be easier to schedule a
restart of ArchivesSpace due to changed indexing rules than to take
down the MAP.  Also, since the Solr documents are quite a bit smaller
than the JSONModel objects they're derived from, it saves on storage.