Add additional ArchivesSpace-specific validations
The fields referencing agencies and series need to be checked by QSA
staff at the time of import.  On the MAP side we'll let these slide
with whatever values the agency puts in, but on the ArchivesSpace side
we'll do the DB checks and make sure everything is in order.

Need a small change to the validation endpoint to make the active
repository available to the validations, since there's no point
linking to a series in a different repository.